requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
